# Conflicts:
#	crypto/txsigner.go
#	crypto/txsigner_eip155.go
#	crypto/txsigner_frontier.go
#	crypto/txsigner_london.go
#	crypto/txsigner_london_test.go
#	server/server.go

+++ command/genesis/params.go
@@ -324,7 +324,7 @@ func (p *genesisParams) initGenesisConfig() error {
 			BaseFee:    command.DefaultGenesisBaseFee,
 		},
 		Params: &chain.Params{
-			ChainID:      int(p.chainID),
+			ChainID:      int64(p.chainID),
 			Forks:        chain.AllForksEnabled,
 			Engine:       p.consensusEngineConfig,
 			BurnContract: map[string]string{},
